Split tracker redux store into document & editor reducers so that only document is affected by undo/redo
