+++ src/store/configureStore.ts
@@ -11,7 +11,7 @@ import projectMiddleware from "./features/project/projectMiddleware";
 import engineMiddleware from "./features/engine/engineMiddleware";
 import spriteMiddleware from "./features/sprite/spriteMiddleware";
 import throttleMiddleware from "./features/throttle/throttleMiddleware";
-import trackerMiddleware from "./features/tracker/trackerMiddleware";
+import trackerDocumentMiddleware from "./features/trackerDocument/trackerDocumentMiddleware";
 
 export type RootState = ReturnType<typeof rootReducer>;
 
@@ -36,7 +36,7 @@ const store = configureStore({
     warningsMiddleware,
     undoMiddleware,
     clipboardMiddleware,
-    trackerMiddleware,
+    trackerDocumentMiddleware,
   ])
 });
