+++ src/store/features/trackerDocument/trackerDocumentActions.ts
@@ -1,0 +1,3 @@
+import { actions } from "./trackerDocumentState";
+
+export default actions;
